+++ graphics/r_window.h
@@ -2,7 +2,7 @@
 //      handles window context creation and management, as well as the window related datatypes
 #pragma once
 
-#include "r_types.h"
+#include "r_api.h"
 #include "fresa_math.h"
 
 namespace fresa::graphics
@@ -14,10 +14,11 @@ namespace fresa::graphics
     //: window type
     //      contains the window and monitor references, as well as its size (not the framebuffer size)
     struct Window {
-        GLFWwindow* window;
-        GLFWmonitor* monitor;
-        Vec2<ui16> size;
+        GLFWwindow* window = nullptr;
+        GLFWmonitor* monitor = nullptr;
+        Vec2<ui16> size = {0, 0};
     };
+    inline std::unique_ptr<const Window> win;
 
     // ···········
     // · SYSTEMS ·
@@ -28,7 +29,7 @@ namespace fresa::graphics
         //: initialize window and callbacks
         Window create();
 
-        //: resize callback (implemented on r_*_api.cpp)
+        //: resize callback (implemented on r_api_*.cpp)
         void onResize(GLFWwindow* window, int width, int height);
 
         //: get current monitor